daemon: split interface-naming startup out of daemon_run.go
Pure code-motion refactor (#5661 cohort). Lifts the startup
interface-naming / enumeration helpers into a sibling file in package
daemon:

  - setupInterfaceNaming
  - namingParamsFromConfig
  - applyStartupNamingForConfig
  - maybeReapplyConfigArrivalNaming
  - runBootstrapExitStartup

Declarations moved verbatim; no rename, no reorder, no logic change.
Only the per-file import set is narrowed.

After this final split daemon_run.go retains just the core lifecycle:
buildRuntimeDataPlane (canary-pinned), Run, and the startup-phase
orchestration (startupSignalContext, startupPhase, runStartupPhases,
runStartupOrAbort, startReconcileRGStateLoop) — 836 lines, down from
~2820, under the ~2000 threshold.

Validation: go build ./..., go test ./pkg/daemon/..., and the
pkg/dataplane retirement-boundary canary all pass; gofmt clean. A
per-declaration byte-identity check confirms all 32 moved/retained
top-level decls are unchanged.
